fix: change omp parallel
